refactor(ui/setup): fold setup-wizard grandchild prop-drilling into setupState
CloudAttachPanel, LocalModelsStatus, ProviderOAuthList and SystemCheckStep now
read the setupState store directly instead of receiving threaded props, mirroring
Screen1ModelsStep/ReviewStep. Screen1ModelsStep and +page.svelte drop the drilled
props (and the now-unused local aliases). Behavior-identical mechanical swaps.

DeployStep is deferred: the store's `deployData` type is narrower than the
`phase`/`imageWarning` fields DeployStep renders, so a clean store-read would
require widening the store's DeployData type — beyond a mechanical swap.
